+++ 404.php
@@ -1,0 +1,14 @@
+<?php get_template_part('parts/header'); ?>
+
+<div id="primary" class="content-area">
+	<main id="main" class="site-main">
+		<section class="error-404 not-found">
+			<h1 class="ku">空</h1>
+			<p><em>La forma è vuoto, il vuoto è forma</em></p>
+			<p class="info">La pagina che stavi cercando non esiste.</p>
+			<p><a href="/">Ritorna alla homepage</a></p>
+		</section>
+	</main>
+</div>
+
+<?php get_template_part('parts/footer'); ?>