+++ patterns/template-home-readings.php
@@ -1,15 +1,5 @@
 
 <section class="home-section <?php if ( ! is_user_logged_in() ) { echo 'czcr-guest-continue'; } ?>" id="readinglist">
-	<?php
-		get_template_part(
-			'parts/cta-title-link',
-			null,
-			[
-				"url"	=> "",
-				"title"	=> "Continua a Leggere"
-			]
-		);
-
-		echo do_shortcode('[readings limit="5"]');
-	?>
+	<h4 class="section-title">Continua a Leggere</h4>
+	<?php echo do_shortcode('[readings limit="5"]'); ?>
 </section>