+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
+- Fix `.success-message` appearance
 - Fix a CSS issue that prevented the correct styling of #footnotes-toggle 
 
 ## [0.8.0] 2025-08-18