+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.9.0] 2025-08-21
 ### Added
 - Add `404.php` page
 - Add `inc/authentication.php` to manage user registration email verification, hide the WordPress top bar and disallow accessing the /dashboard to non admin users.
@@ -137,7 +137,8 @@ the slug's name
 - Initial `index.php`, `style.css`, and `functions.php`
 - This Changelog
 
-[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.9.0
 [0.8.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.8.0
 [0.7.1]: https://github.com/erremauro/cigno-zen/releases/tag/v0.7.1
 [0.7.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.7.0